Update linal version.

+++ src/include/Geometry/Distance/DistancePlane.hpp
@@ -9,14 +9,14 @@
 namespace Geometry
 {
 
-template <typename T, std::size_t D>
+template <typename T, std::uint8_t D>
 GEO_NODISCARD constexpr T distance(const Plane<T>& plane, linal::vec<T, D> vec) noexcept
 {
   linal::vec<T, D> vecToPlane = plane.get_origin() - vec;
   return std::abs(linal::dot(vecToPlane, plane.get_normal()));
 }
 
-template <typename T, std::size_t D>
+template <typename T, std::uint8_t D>
 GEO_NODISCARD constexpr T distance(linal::vec<T, D> vec, const Plane<T>& plane) noexcept
 {
   return distance(plane, vec);
